+++ src/components/definition-list/definition-list-test.stories.tsx
@@ -5,10 +5,13 @@ import Dd from "../definition-list/dd.component";
 import Typography from "../typography";
 import Hr from "../hr";
 import Box from "../box";
+import Tile from "../tile";
+import Icon from "../icon";
+import Pill from "../pill";
 
 export default {
   title: "Definition-list/Test",
-  includeStories: ["DefaultStory"],
+  includeStories: ["DefaultStory", "UsingBoxToOverrideBackgroundColor"],
   parameters: {
     info: { disable: true },
     chromatic: {
@@ -17,16 +20,9 @@ export default {
   },
 };
 
-export const DefaultStory = ({
-  children,
-  w,
-  dtTextAlign,
-  ddTextAlign,
-  asSingleColumn,
-  ...args
-}: DlProps) => {
+export const DefaultStory = (props: DlProps) => {
   return (
-    <Dl {...args}>
+    <Dl {...props}>
       <Dt>First</Dt>
       <Dd>Description</Dd>
       <Dt>Second</Dt>
@@ -88,4 +84,94 @@ export const DLBoxComponent = () => {
       </Box>
     </div>
   );
+};
+
+export const UsingBoxToOverrideBackgroundColor = () => (
+  <Tile width="368px">
+    <Dl m={0} dtTextAlign="left" ddTextAlign="right">
+      <Dt>
+        <Typography variant="segment-header" as="h2">
+          Foo
+        </Typography>
+      </Dt>
+      <Dd>Bar</Dd>
+      <Box
+        backgroundColor="yellow"
+        display="flex"
+        gridColumn="1 / -1"
+        alignItems="center"
+        py={2}
+        justifyContent="space-between"
+        data-element="box1"
+      >
+        <Dt mb={0}>
+          <Box display="inline-flex" alignItems="center">
+            <Icon type="tick" mr={1} />
+            <Box>Foo</Box>
+          </Box>
+        </Dt>
+        <Dd mb={0}>Bar</Dd>
+      </Box>
+      <Dt>
+        <Pill>Foo</Pill>
+      </Dt>
+      <Dt>
+        <Box display="inline-flex" alignItems="center">
+          <Icon type="tick" mr={1} />
+          <Box>Foo</Box>
+        </Box>
+      </Dt>
+      <Dd>Bar</Dd>
+      <Dt>
+        <Box display="inline-flex" alignItems="center">
+          <Icon type="tick" mr={1} />
+          <Box>Foo</Box>
+        </Box>
+      </Dt>
+      <Dd>Bar</Dd>
+      <Dt>
+        <Pill>Foo</Pill>
+      </Dt>
+      <Dt>
+        <Box display="inline-flex" alignItems="center">
+          <Icon type="tick" mr={1} />
+          <Box>Foo</Box>
+        </Box>
+      </Dt>
+      <Dd>Bar</Dd>
+      <Box
+        backgroundColor="greenyellow"
+        display="flex"
+        gridColumn="1 / -1"
+        alignItems="baseline"
+        py={2}
+        justifyContent="space-between"
+        data-element="box2"
+      >
+        <Dt mb={0}>
+          Foo
+          <Typography mb={0}>(foo)</Typography>
+        </Dt>
+        <Dd>Bar</Dd>
+      </Box>
+      <Box display="flex" gridColumn="1 / -1" justifyContent="space-between">
+        <Dt>Foo</Dt>
+        <Dd>Bar</Dd>
+      </Box>
+      <Dt>Foo</Dt>
+      <Dd>Bar</Dd>
+      <Dt mb={0}>
+        <Typography variant="h3">Foo</Typography>
+        <Typography mb={0}>(foo)</Typography>
+      </Dt>
+      <Dd mb={0}>
+        <Typography mb={0}>Bar</Typography>
+        <Typography mb={0}>Bar</Typography>
+      </Dd>
+    </Dl>
+  </Tile>
+);
+
+UsingBoxToOverrideBackgroundColor.parameters = {
+  chromatic: { disableSnapshot: false },
 };